consolidating items into UI
