+++ KlockWork/Views/Entities/Tasks/TaskItem.swift
@@ -10,6 +10,7 @@ import SwiftUI
 import KWCore
 
 struct TaskItem: View, Identifiable {
+    typealias UI = WidgetLibrary.UI
     @EnvironmentObject public var state: Navigation
     public var id: UUID = UUID()
     public let task: LogTask
@@ -76,9 +77,9 @@ struct TaskItem: View, Identifiable {
                     Spacer()
                     if self.isHighlighted {
                         VStack(alignment: .trailing, spacing: 0) {
-                            RowActionButton(callback: self.actionOnTaskCancel, icon: "calendar.badge.minus", helpText: "Cancel task", highlightedColour: .red)
-                            RowActionButton(callback: self.actionOnTaskDelay, icon: "clock.fill", helpText: "Delay task 1 day", highlightedColour: .yellow)
-                            RowActionButton(callback: self.actionOnTaskComplete, icon: "checkmark.circle.fill", helpText: "Task complete!", highlightedColour: .green)
+                            UI.RowActionButton(callback: self.actionOnTaskCancel, icon: "calendar.badge.minus", helpText: "Cancel task", highlightedColour: .red)
+                            UI.RowActionButton(callback: self.actionOnTaskDelay, icon: "clock.fill", helpText: "Delay task 1 day", highlightedColour: .yellow)
+                            UI.RowActionButton(callback: self.actionOnTaskComplete, icon: "checkmark.circle.fill", helpText: "Task complete!", highlightedColour: .green)
                         }
                         .frame(width: 30)
                     }
